+++ Sources/SwiftUIKit/Lists/ListButtonGroupStyle.swift
@@ -9,8 +9,7 @@
 #if os(iOS)
 import SwiftUI
 
-/// This button style can generate inline button groups that mimics the look of the
-/// topmost iOS Contact button group.
+@available(*, deprecated, renamed: "ListActionButtonGroup")
 public struct ListButtonGroupStyle: ButtonStyle {
     
     /// Create a custom form group button style.